marketing_ai_crew/requirements.txt
crewai>=0.80.0
crewai-tools>=0.17.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
duckduckgo-search>=6.3.0
flask>=3.1.0
flask-cors>=5.0.0
python-dotenv>=1.0.1
pydantic>=2.10.0
rich>=13.9.0
click>=8.1.8
requests>=2.32.0
